Model Loader Progress
I have made significant progress to the model loader, I have created a system in which a model is loaded into the scene.

The next step is to work out the numerous bugs that ocur during the loading process, (The shape isn't properly loaded into the scene).

I need to make a basic solution first then clean it up and abstract it once it works.